requirements.txt
fastapi
pillow
websocket-client
uvicorn
jinja2
reportlab
python-multipart
pymongo
dotenv
email-validator
PyMuPDF